Als 4662 (#122)
* Update header name

Previously, our header could not be found due to case sensitivity. Our approach should be case in sensitivity.

* Add approach to printing the QueryRequest

There is currently no toString for the queryRequest object. Printing the query itself requires converting the json to a string. This is done with a utility method instead of a toString. I did this to avoid having an ObjectMapper in the QueryRequest class.

* Add additional information to diagnose header issue

* Create RequestScope header context

Previously, the headers were created during initialization as a singleton. This means that the Headers were never updated after initialization leading to them always being null. To avoid this I created a Header class that is application scoped and can be injected into each class. This seemed like a better approach than passing the header context from each request method to the necessary class methods.

* Fix unit test by mocking HeaderContext

* Change one request method to use HttpHeaders

Changed one request method to use the @Context HttpHeaders, so I can test if the value headers in the param context are also null.

* Removed HeaderContext file and injections

This approach does not work and the headers are still null.

* Update all methods to pass header context

* Update code based on review feedback