Sync new Release/0008 with Master (#168)
* [ALS-5422] Add persistence.xml to visualization resource (#161)

The visualization resource is failing to start due to an error injecting persistence unit into CDI managed bean. It is unable to find a persistence unit named ''.

* [ALS-0000] Fix aggregate resource

- Very confused as to how this was working before

* [ALS-5422] Add persistence.xml to visualization resource (#161) (#162) (#163)

The visualization resource is failing to start due to an error injecting persistence unit into CDI managed bean. It is unable to find a persistence unit named ''.

---------

Co-authored-by: Luke Sikina <lucas.skina@gmail.com>